♻️ refactor: all owner methods 100% working

+++ src/main/java/br/com/petcolumbia/api_pet_columbia/core/application/usecase/owner/DeleteOwnerByIdUseCase.java
@@ -1,6 +1,7 @@
 package br.com.petcolumbia.api_pet_columbia.core.application.usecase.owner;
 
-import br.com.petcolumbia.api_pet_columbia.core.adapter.OwnerGateway;
+import br.com.petcolumbia.api_pet_columbia.core.adapter.owner.OwnerGateway;
+import br.com.petcolumbia.api_pet_columbia.core.application.exception.EntityNotFoundException;
 
 public class DeleteOwnerByIdUseCase {
     private final OwnerGateway ownerGateway;
@@ -10,6 +11,9 @@ public class DeleteOwnerByIdUseCase {
     }
 
     public void execute(Integer id) {
+        if (ownerGateway.existsById(id))
+            throw new EntityNotFoundException("Não encontrado usuário com id:" + id);
+        
         ownerGateway.deleteOwnerById(id);
     }
 }
